Revert 65e51c84: filtering in escalus_tcp
This was not a well thought out idea.
The functionality this changeset intended to make feasible was
achieved in a simpler way from within the relevant testsuite.

OTOH if the change is desirable the code remains in the history.
Keep in mind, though, that support for filtering should be added to all
the existing transports in parallel.