Support optional handlers for received stanzas
When get_stanza_* is used with a predicate,
the received stanzas not matching the predicate are skipped.

However, you can put a list of handlers in client properties:

  {received_stanza_handlers, [Handler1, Handler2, ...]}

This way instead of ignoring the stanza,
the handlers will be applied to it sequentially
until on of them returns 'true'.

This can be used for:
  - updating metrics,
  - handling requests (e.g. pings) from the server,
  - responding to incoming messages,
  - logging unexpected stanzas
etc.