Implement PLUS variants in escalus
With this changes, if using any of the scram mechanisms, escalus will
always send a `gs2-cbind-flag` set to `y`, unless explicitly configured
to use channel binding, in which case it will send such flag as
`p=tls-unique`, regardless of what the server advertised. This needs to
be taken into account when testing, as we could be mismatching them,
with the following scenarios failing auth:

- Server advertises `PLUS`, client sends a `y=`: it means that the
client thinks that the server does not support channel binding while the
client does so, which could mean that the advertised features could have
been intercepted do avoid channel binding, i.e., a MitM or a downgrade
attack, hence the server must fail this authentication.

- Server does not advertise `PLUS`, client sends a `p=tls-unique`: it
means that the client is requesting a method that the server does
obviously not support, hence the server will also fail this
authentication.

Care must be taken to what is intended when using this new feature.